add face models
added models to config dropdown

fixed downloading

updated tests

use hf for face models

formatting
